Factor out FieldInfo
